requirements.txt
Flask==2.1.2
google-generativeai==0.1.0
faiss-cpu==1.7.2
pandas==1.3.3
sentence-transformers==2.1.0
numpy==1.21.2
tqdm==4.62.3
pyngrok==5.1.0
huggingface_hub